testingm

+++ app/controllers/clubs_controller.rb
@@ -1,25 +1,24 @@
 require 'pry'
 class ClubsController < ApplicationController
   
+ 
+    @@players = Player.all
+    @@notgk=Player.select {|p| p["position"] != "GK"}.sort_by{|p| p.position}
+    @@forward = Player.select {|p| p["position"] == "Fwd"}
+    @@midfield = Player.select {|p| p["position"] == "Mid"}
+    @@defender = Player.select {|p| p["position"] == "Def"}
+    @@goalie = Player.select {|p| p["position"] == "GK"}
   # GET: /players
   get "/clubs" do
-    
-    if logged_in?
     @clubs = Club.all.reverse
     erb :"/clubs/index.html"
-    else
-      redirect '/login'
-    end
   end
-
+    
+    
   # GET: /players/new
   get "/clubs/new" do
-    require_login
-      @players=Player.select {|p| p["position"] != "GK"}
-      @fwd = Player.select {|p| p["position"] == "Fwd"}
-      @mid = Player.select {|p| p["position"] == "Mid"}
-      @def = Player.select {|p| p["position"] == "Def"}
-      @gk = Player.select {|p| p["position"] == "GK"}
+    @@players
+     
       erb :"clubs/new.html"
     
   end
@@ -27,7 +26,7 @@ class ClubsController < ApplicationController
   # POST: /players
   post "/clubs" do
     params["club"][:user_id] = session[:user_id]
-    club = Club.create(params["club"])
+    club = Club.new(params["club"])
     if club.save
       redirect "/clubs/#{club.id}"  
     else
@@ -41,7 +40,7 @@ class ClubsController < ApplicationController
   # GET: /players/5
   get "/clubs/:id" do
     
-    @club = Club.find_by_id(params[:id])
+   if @club = Club.find_by_id(params[:id])
     @any = Player.find_by_id(@club.any_id)
     @fwd = Player.find_by_id(@club.fwd_id)
      
@@ -50,11 +49,15 @@ class ClubsController < ApplicationController
     @gk = Player.find_by_id(@club.gk_id)
     @user = User.find_by_id(@club.user_id)
     erb :"/clubs/show.html"
+   else
+      redirect '/clubs'
+   end
   end
 
   # GET: /players/5/edit
   get "/clubs/:id/edit" do
     @club = Club.find(params["id"])
+    if @club.user_id == session[:user_id] 
     @any = Player.find_by_id(@club.any_id)
     
    @fwd = Player.find_by_id(@club.fwd_id)
@@ -62,12 +65,15 @@ class ClubsController < ApplicationController
    @def = Player.find_by_id(@club.def_id)
    @gk = Player.find_by_id(@club.gk_id)
    @user = User.find_by_id(@club.user_id)
-   @players=Player.select {|p| p["position"] != "GK"}
+    @notgk=Player.select {|p| p["position"] != "GK"}
     @forward = Player.select {|p| p["position"] == "Fwd"}
     @midfield = Player.select {|p| p["position"] == "Mid"}
     @defender = Player.select {|p| p["position"] == "Def"}
     @goalie = Player.select {|p| p["position"] == "GK"}
     erb :"/clubs/edit.html"
+    else
+      redirect'/clubs'
+    end
   end
 
   # PATCH: /players/5
@@ -90,8 +96,14 @@ class ClubsController < ApplicationController
 
   # DELETE: /players/5/delete
   delete "/clubs/:id" do
-      club = Club.find(params[:id])
+    club = Club.find(params[:id])
+    if club.user_id == session[:user_id] 
       club.destroy
+    end
     redirect "/clubs"
   end
+    
 end
+
+     
+
